+++ views/auth-index.pug
@@ -17,7 +17,7 @@ block content
           div(class='sub-navbar') 
             button(class='snb snButton1') Following 
             button(class='snbOff snButton2') Recommended For You
-          div(class='reco-content')
+          div(class='reco-content pContainer')
             each posts in post
               +detailedStoryGridItem(posts)
           div(class='follow-content')
@@ -29,3 +29,4 @@ block content
         div(class='e-c-2')
             div(id='eop-div')
             include includes/fixed-footer.pug
+  footer.bott
